+++ src/error/parse_from_description.rs
@@ -1,6 +1,9 @@
 //! Error parsing an input into a [`Parsed`](crate::parsing::Parsed) struct
 
+use core::convert::TryFrom;
 use core::fmt;
+
+use crate::error;
 
 /// An error that occurred while parsing the input into a [`Parsed`](crate::parsing::Parsed) struct.
 #[cfg_attr(__time_03_docs, doc(cfg(feature = "parsing")))]
@@ -33,5 +36,16 @@ impl std::error::Error for ParseFromDescription {}
 impl From<ParseFromDescription> for crate::Error {
     fn from(original: ParseFromDescription) -> Self {
         Self::ParseFromDescription(original)
+    }
+}
+
+impl TryFrom<crate::Error> for ParseFromDescription {
+    type Error = error::DifferentVariant;
+
+    fn try_from(err: crate::Error) -> Result<Self, Self::Error> {
+        match err {
+            crate::Error::ParseFromDescription(err) => Ok(err),
+            _ => Err(error::DifferentVariant),
+        }
     }
 }